aleph/backend/requirements.txt
fastapi>=0.110,<0.120
uvicorn[standard]>=0.30
sse-starlette>=2.1
psycopg[binary]>=3.2
psycopg-pool>=3.2
pgvector>=0.3
numpy>=1.26
umap-learn>=0.5.5
hdbscan>=0.8.38
python-dotenv>=1.0
pydantic>=2.6
# transitive deps from the MCP memory package (imported via MCP_PATH)
tenacity>=8.2
google-genai>=0.3
# tests
pytest>=8
pytest-asyncio>=0.23
httpx>=0.27
